requirements.txt
asgiref==3.2.10
botocore==1.18.16
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.4.7
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.44.0
django-ckeditor==6.0.0
django-crispy==1.8.1
django-crispy-forms==1.11.0
django-environ==0.4.5
django-js-asset==1.2.2
django-markdown==0.8.4
django-taggit==1.2.0
djangorestframework==3.12.4
google-auth==1.22.1
googleapis-common-protos==1.52.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
Markdown==3.3.4
oauthlib==3.1.0
Pillow==8.2.0
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.25.10
whitenoise==4.1.2
wrapt==1.12.1
